Split `escapeJS` in `linker` and `js-envs`.
And specialize them to the particular needs of each subproject:

* `java.lang.StringBuilder` versus `java.io.Writer` instead of the
  generic `java.lang.Appendable`.
* With or without tracking of how many characters are written out.

Also review and specialize the one in `ir`.
