+++ sbt-plugin/src/main/scala/org/scalajs/sbtplugin/ScalaJSPluginInternal.scala
@@ -29,7 +29,6 @@ import sbt.complete.DefaultParsers._
 import org.portablescala.sbtplatformdeps.PlatformDepsPlugin.autoImport._
 
 import org.scalajs.io._
-import org.scalajs.io.JSUtils.escapeJS
 
 import org.scalajs.linker._
 import org.scalajs.linker.irio._
