+++ test-adapter/src/main/scala/org/scalajs/testing/adapter/TestAdapter.scala
@@ -17,10 +17,10 @@ import scala.concurrent.duration._
 import scala.collection.concurrent.TrieMap
 
 import org.scalajs.io._
-import org.scalajs.io.JSUtils.escapeJS
 import org.scalajs.logging._
 
 import org.scalajs.jsenv._
+import org.scalajs.jsenv.JSUtils.escapeJS
 
 import org.scalajs.testing.common._
